[Chore]: Corrected the type hinting value

+++ 0x02-python_async_comprehension/1-async_comprehension.py
@@ -5,7 +5,7 @@ from typing import List
 async_generator = __import__("0-async_generator").async_generator
 
 
-async def async_comprehension() -> List[int]:
+async def async_comprehension() -> List[float]:
     """
         Function that generates 10 different numbers using async comprehension
         over async generation
